requirements.txt
beautifulsoup4
Flask
requests
python-dotenv